requirements.txt
numpy~=1.21.5
torch~=1.12.1
scikit-learn~=1.0.2
ClusterContrast~=1.0.0
torchvision~=0.13.1
tensorboardX~=2.2
setuptools~=65.6.3
Pillow~=9.4.0
faiss~=1.7.3
scipy~=1.7.3
infomap~=2.7.1
tqdm~=4.66.4
timm~=0.9.12